Remove S3 references from ShareObjectService, split sharelogs into its own service
