+++ backend/dataall/modules/s3_datasets_shares/services/share_processors/glue_table_share_processor.py
@@ -13,7 +13,7 @@ from dataall.modules.s3_datasets.db.dataset_models import DatasetTable
 from dataall.modules.shares_base.services.share_exceptions import PrincipalRoleNotFound
 from dataall.modules.s3_datasets_shares.services.share_managers import LFShareManager
 from dataall.modules.s3_datasets_shares.aws.ram_client import RamClient
-from dataall.modules.s3_datasets_shares.services.share_object_service import ShareObjectService
+from dataall.modules.shares_base.services.share_object_service import ShareObjectService
 from dataall.modules.s3_datasets_shares.services.share_item_service import ShareItemService
 from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
 from dataall.modules.shares_base.db.share_object_state_machines import ShareItemSM
